move set seed to core
